+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ##### mkdoc 1.1.0 - (28 September 2015)
 
-    * Add --check to check for orphan images
-    * Add --noop to skip executing pandoc
+    * Add --check [img] to check for orphan images
+    * Add --noop to skip pandoc execution
     * Add --renumber to renumber markdown files
+    * Add {.Day} {.Month} {.Year} {.Hour} {.Minute} {.Second} metadata
 
 ##### mkdoc 1.0.0 - (19 September 2015)